requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
CacheControl==0.12.6
cached-property==1.5.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
daphne==2.5.0
distlib==0.3.1
Django==3.1.1
django-celery==3.3.1
django-celery-results==1.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-environ==0.6.0
django-mathfilters==1.0.0
django-tinymce==3.3.0
djangorestframework==3.12.4
dnspython==1.16.0
docutils==0.16
et-xmlfile==1.0.1
eventlet==0.28.0
feedparser==5.2.1
filelock==3.0.12
fontawesome==5.10.1.post1
gevent==20.9.0
greenlet==0.4.17
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==5.1.2
incremental==17.5.0
isort==5.4.2
jdcal==1.4.1
lastversion==1.1.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
mccabe==0.6.1
monotonic==1.5
msgpack==1.0.0
mutapath==0.16.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.4
path==15.0.0
pefile==2019.4.18
Pillow==8.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
redis==3.5.3
remotezip==0.9.2
requests==2.24.0
semver==2.10.2
service-identity==18.1.0
singletons==0.2.5
six==1.15.0
soupsieve==2.0.1
spark-parser==1.8.9
sqlparse==0.3.1
tabulate==0.8.7
toml==0.10.1
tqdm==4.48.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typeshed-client==0.4.1
typing-extensions==3.7.4.3
uncompyle6==3.7.4
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
wrapt==1.12.1
xdis==5.0.13
xlrd==1.2.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.2